# mobile/lib/swagger/api.json
{"openapi":"3.0.0","paths":{"/api/game-template":{"get":{"operationId":"GameTemplateController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GameTemplate"}}}}}},"tags":["Game Template"]},"post":{"operationId":"GameTemplateController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServerCreateGameDto"}}}},"responses":{"201":{"description":""},"400":{"description":"Dto contains invalid field","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"401":{"description":"Missing or bad JWT","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Game Template"],"security":[{"bearer":[]}]},"delete":{"operationId":"GameTemplateController_removeAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Game Template"]}},"/api/game-template/length":{"get":{"operationId":"GameTemplateController_getLength","summary":"Get the number of game templates that exist","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"number"}}}}},"tags":["Game Template"]}},"/api/game-template/{id}":{"get":{"operationId":"GameTemplateController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameTemplate"}}}},"400":{"description":"Invalid ObjectId","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"404":{"description":"Game template not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Game Template"]},"delete":{"operationId":"GameTemplateController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""},"400":{"description":"Invalid ObjectId","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"404":{"description":"Game template not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Game Template"]}},"/api/game-template/page/{page}":{"get":{"operationId":"GameTemplateController_findPage","parameters":[{"name":"page","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GameTemplate"}}}}},"400":{"description":"Invalid page number, should be greater or equal to zero","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Game Template"]}},"/api/images-differences":{"post":{"operationId":"ImagesDifferencesController_createDiffImage","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateImagesDifferencesDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServerCreateDiffResult"}}}},"400":{"description":"The base64 encoded images should be in bmp, jpg or png format","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Images Differences"]}},"/api/auth/login":{"post":{"operationId":"AuthController_signIn","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JwtTokenDto"}}}},"400":{"description":"User already connected","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"401":{"description":"Password is invalid","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"404":{"description":"User does not exist","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Auth"]}},"/api/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JwtTokenDto"}}}},"401":{"description":"Password is invalid","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"404":{"description":"User does not exist","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"409":{"description":"User already exists or User already connected","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Auth"]}},"/api/auth/refresh-jwt":{"get":{"operationId":"AuthController_refreshJwt","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JwtTokenDto"}}}},"401":{"description":"Missing or bad JWT","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/api/auth/changePassword":{"post":{"operationId":"AuthController_changePassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordDto"}}}},"responses":{"200":{"description":""},"401":{"description":"Missing or bad JWT","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"404":{"description":"User does not exist","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/api/user":{"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Users"]}},"/api/user/profile":{"get":{"operationId":"UsersController_getProfile","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"401":{"description":"Missing or bad JWT","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Users"],"security":[{"bearer":[]}]}},"/api/user/theme":{"get":{"operationId":"UsersController_findTheme","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Theme"}}}},"401":{"description":"Missing or bad JWT","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Users"],"security":[{"bearer":[]}]},"patch":{"operationId":"UsersController_updateTheme","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Theme"}}}},"responses":{"200":{"description":""},"401":{"description":"Missing or bad JWT","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Users"],"security":[{"bearer":[]}]}},"/api/user/stats/{username}":{"get":{"operationId":"UsersController_getStatistics","parameters":[{"name":"username","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"number"}}}}},"401":{"description":"Missing or bad JWT","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Users"],"security":[{"bearer":[]}]}},"/api/user/balance":{"get":{"operationId":"UsersController_getBalance","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"number"}}}},"401":{"description":"Missing or bad JWT","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Users"],"security":[{"bearer":[]}]}},"/api/user/stats":{"patch":{"operationId":"UsersController_updateDifferencesFoundPercentage","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DifferenceFoundRatioDto"}}}},"responses":{"200":{"description":""},"401":{"description":"Missing or bad JWT","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Users"],"security":[{"bearer":[]}]}},"/api/user/replays":{"get":{"operationId":"UsersController_getReplays","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReceiveReplayDto"}}}}},"401":{"description":"Missing or bad JWT","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Users"],"security":[{"bearer":[]}]}},"/api/user/addReplay":{"post":{"operationId":"UsersController_addReplay","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendReplayDto"}}}},"responses":{"201":{"description":""},"401":{"description":"Missing or bad JWT","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Users"],"security":[{"bearer":[]}]}},"/api/user/deleteReplay/{replayId}":{"delete":{"operationId":"UsersController_deleteReplay","parameters":[{"name":"replayId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""},"401":{"description":"Missing or bad JWT","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Users"],"security":[{"bearer":[]}]}},"/api/user/locale":{"get":{"operationId":"UsersController_findLocale","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocaleDto"}}}},"401":{"description":"Missing or bad JWT","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Users"],"security":[{"bearer":[]}]},"patch":{"operationId":"UsersController_updateLocale","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocaleDto"}}}},"responses":{"200":{"description":""},"401":{"description":"Missing or bad JWT","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Users"],"security":[{"bearer":[]}]}},"/api/user/username":{"patch":{"operationId":"UsersController_updateUsername","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsernameDto"}}}},"responses":{"200":{"description":""},"401":{"description":"Missing or bad JWT","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"409":{"description":"User already exists","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Users"],"security":[{"bearer":[]}]}},"/api/user/change-avatar":{"patch":{"operationId":"UsersController_changeAvatar","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangeAvatarDto"}}}},"responses":{"200":{"description":""},"401":{"description":"Missing or bad JWT","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Users"],"security":[{"bearer":[]}]}},"/api/history":{"get":{"operationId":"HistoryController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/History"}}}}}},"tags":["History"]},"delete":{"operationId":"HistoryController_removeAll","parameters":[],"responses":{"200":{"description":""}},"tags":["History"]}},"/api/history/{username}":{"get":{"operationId":"HistoryController_findByUsername","parameters":[{"name":"username","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/History"}}}}}},"tags":["History"]}},"/api/chat/available-chats":{"get":{"operationId":"ChatController_getAvailableChats","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ChatSchema"}}}}},"401":{"description":"Missing or bad JWT","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Chat"],"security":[{"bearer":[]}]}},"/api/chat/joined-chats":{"get":{"operationId":"ChatController_getJoinedChats","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ChatIdSchema"}}}}},"401":{"description":"Missing or bad JWT","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Chat"],"security":[{"bearer":[]}]}},"/api/chat/general-chat-id":{"get":{"operationId":"ChatController_getGeneralChatId","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatIdSchema"}}}},"401":{"description":"Missing or bad JWT","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Chat"],"security":[{"bearer":[]}]}},"/api/friends/friend-list":{"get":{"operationId":"FriendsController_getFriendList","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}},"401":{"description":"Missing or bad JWT","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Friends"],"security":[{"bearer":[]}]}},"/api/friends/friend-list/{username}":{"get":{"operationId":"FriendsController_getFriendListByUsername","parameters":[{"name":"username","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}},"401":{"description":"Missing or bad JWT","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Friends"],"security":[{"bearer":[]}]}},"/api/friends/outgoing-request-list":{"get":{"operationId":"FriendsController_getOutgoingRequestList","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}},"401":{"description":"Missing or bad JWT","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Friends"],"security":[{"bearer":[]}]}},"/api/friends/incoming-request-list":{"get":{"operationId":"FriendsController_getIncomingRequestList","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}},"401":{"description":"Missing or bad JWT","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Friends"],"security":[{"bearer":[]}]}},"/api/account-history":{"get":{"operationId":"AccountHistoryController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AccountHistory"}}}}}},"tags":["Account History"]},"delete":{"operationId":"AccountHistoryController_removeAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Account History"]}},"/api/account-history/{username}":{"get":{"operationId":"AccountHistoryController_findByUsername","parameters":[{"name":"username","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AccountHistory"}}}}}},"tags":["Account History"]}},"/api/market/buyGame":{"post":{"operationId":"MarketController_buyGame","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuyGameDto"}}}},"responses":{"200":{"description":""},"400":{"description":"Dto contains invalid field","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"401":{"description":"Missing or bad JWT","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Market"],"security":[{"bearer":[]}]}},"/api/market/buyableGames":{"get":{"operationId":"MarketController_getBuyableGames","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GameTemplate"}}}}},"401":{"description":"Missing or bad JWT","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Market"],"security":[{"bearer":[]}]}},"/api/market/availableGames":{"get":{"operationId":"MarketController_getAvailableGames","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GameTemplate"}}}}},"401":{"description":"Missing or bad JWT","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Market"],"security":[{"bearer":[]}]}}},"info":{"title":"Cadriciel Serveur","description":"Serveur du projet de base pour le cours de LOG3900","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"GameTemplate":{"type":"object","properties":{"_id":{"type":"string"},"name":{"type":"string"},"deleted":{"type":"boolean"},"difficulty":{"enum":["easy","hard"],"type":"string"},"firstImage":{"type":"string"},"secondImage":{"type":"string"},"price":{"type":"number"},"creator":{"type":"string"},"nGroups":{"type":"number"}},"required":["_id","name","deleted","difficulty","firstImage","secondImage","price","creator","nGroups"]},"ServerCreateGameDto":{"type":"object","properties":{"name":{"type":"string","maxLength":256},"firstImage":{"type":"string"},"secondImage":{"type":"string"},"radius":{"type":"number","minimum":0,"maximum":15},"price":{"type":"number","minimum":0}},"required":["name","firstImage","secondImage","radius","price"]},"ErrorResponseDto":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"enum":["Missing or bad JWT","Cannot Rename Admin","User already connected","User does not exist","User already exists","Password is invalid","Game template not found","Invalid page number, should be greater or equal to zero","Invalid ObjectId","The base64 encoded images should be in bmp, jpg or png format","Image not found","Dto contains invalid field","Not enough money"],"type":"string"},"error":{"type":"string"}},"required":["statusCode","message","error"]},"CreateImagesDifferencesDto":{"type":"object","properties":{"image1Base64":{"type":"string"},"image2Base64":{"type":"string"},"radius":{"type":"number"}},"required":["image1Base64","image2Base64","radius"]},"ServerCreateDiffResult":{"type":"object","properties":{"nGroups":{"type":"number"},"difficulty":{"type":"string","enum":["easy","hard"]},"diffImage":{"type":"string"}},"required":["nGroups","difficulty","diffImage"]},"LoginDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]},"JwtTokenDto":{"type":"object","properties":{"accessToken":{"type":"string"}},"required":["accessToken"]},"CreateUserDto":{"type":"object","properties":{"username":{"type":"string"},"avatarUrl":{"type":"string"},"password":{"type":"string"}},"required":["username","avatarUrl","password"]},"ChangePasswordDto":{"type":"object","properties":{"password":{"type":"string"}},"required":["password"]},"UserDto":{"type":"object","properties":{"avatarUrl":{"type":"string"},"username":{"type":"string"}},"required":["avatarUrl","username"]},"Theme":{"type":"object","properties":{"primaryColor":{"type":"string"}},"required":["primaryColor"]},"DifferenceFoundRatioDto":{"type":"object","properties":{"differenceFound":{"type":"number"}},"required":["differenceFound"]},"ReplayEvent":{"type":"object","properties":{"name":{"type":"string"},"time":{"type":"number"},"data":{"type":"object"}},"required":["name","time","data"]},"ReceiveReplayDto":{"type":"object","properties":{"id":{"type":"string"},"events":{"type":"array","items":{"$ref":"#/components/schemas/ReplayEvent"}},"gameTemplateId":{"type":"string"},"gameMode":{"type":"string","enum":["teamClassic","timeLimitSingleDiff","timeLimitAugmented","timeLimitTurnByTurn","classicMultiplayer"]},"username":{"type":"string"}},"required":["id","events","gameTemplateId","gameMode","username"]},"SendReplayDto":{"type":"object","properties":{"events":{"type":"array","items":{"$ref":"#/components/schemas/ReplayEvent"}},"gameTemplateId":{"type":"string"},"gameMode":{"type":"string","enum":["teamClassic","timeLimitSingleDiff","timeLimitAugmented","timeLimitTurnByTurn","classicMultiplayer"]}},"required":["events","gameTemplateId","gameMode"]},"LocaleDto":{"type":"object","properties":{"locale":{"type":"string"}},"required":["locale"]},"UsernameDto":{"type":"object","properties":{"username":{"type":"string"}},"required":["username"]},"ChangeAvatarDto":{"type":"object","properties":{"avatarUrl":{"type":"string"}},"required":["avatarUrl"]},"History":{"type":"object","properties":{"startTime":{"type":"number"},"totalTime":{"type":"number"},"gameMode":{"enum":["teamClassic","timeLimitSingleDiff","timeLimitAugmented","timeLimitTurnByTurn","classicMultiplayer"],"type":"string"},"winners":{"type":"array","items":{"type":"string"}},"losers":{"type":"array","items":{"type":"string"}},"quitters":{"type":"array","items":{"type":"string"}}},"required":["startTime","totalTime","gameMode","winners","losers","quitters"]},"ChatSchema":{"type":"object","properties":{"chatId":{"type":"string"},"name":{"type":"string"},"creator":{"type":"string"}},"required":["chatId","name","creator"]},"ChatIdSchema":{"type":"object","properties":{"chatId":{"type":"string"}},"required":["chatId"]},"AccountHistory":{"type":"object","properties":{"username":{"type":"string"},"action":{"type":"string"},"timestamp":{"type":"number"}},"required":["username","action","timestamp"]},"BuyGameDto":{"type":"object","properties":{"gameTemplateId":{"type":"string"}},"required":["gameTemplateId"]}}}}